Renamed: components to have prefix Dxp before the names
