+++ src/components/DxpProductIdentifier.vue
