+++ src/components/index.ts
@@ -17,9 +17,9 @@ import '@ionic/vue/css/display.css';
 export { default as DxpImage } from './DxpImage.vue';
 export { default as DxpUserProfile } from './DxpUserProfile.vue'
 export { default as DxpAppVersionInfo } from './DxpAppVersionInfo.vue';
-export { default as LanguageSwitcher } from './LanguageSwitcher.vue';
+export { default as DxpLanguageSwitcher } from './DxpLanguageSwitcher.vue';
 export { default as DxpMenuFooterNavigation } from './DxpMenuFooterNavigation.vue';
 export { default as DxpLogin } from './DxpLogin.vue';
-export { default as OmsInstanceNavigator } from './OmsInstanceNavigator.vue'
-export { default as ProductIdentifier } from "./ProductIdentifier.vue";
-export { default as ShopifyImg } from './ShopifyImg.vue';
+export { default as DxpOmsInstanceNavigator } from './DxpOmsInstanceNavigator.vue'
+export { default as DxpProductIdentifier } from "./DxpProductIdentifier.vue";
+export { default as DxpShopifyImg } from './DxpShopifyImg.vue';
